API change: merger_context is gone
Merger context was removed from the API.

A source and a merger don't more reallocate tuples: a source / a merger
did create a new tuple if stored tuple format differs from acquired one.
The idea was to ensure comparisons are fast. This however had its cost:
we were need to re-create a tuple with another format. Now a source / a
merger don't do that, but a user should ensure tuples has needed
offsets.

Added recommendations how to ensure that comparisons in a merge process
will be fast.